# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCare Hub API
  version: 1.0.0
  description: >
    API para gerenciamento de clientes, veículos, serviços, peças, ordens de
    serviço, orçamentos, aprovações e acompanhamento de status de uma oficina
    mecânica.
servers:
  - url: http://localhost:8080
    description: Ambiente local
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/customers/{customerId}/service-orders:
    get:
      tags:
        - Service Orders
      summary: Lista ordens de serviço de um cliente
      operationId: listServiceOrdersByCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerIdPathParam'
      responses:
        '200':
          description: Ordens de serviço encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/service-orders:
    post:
      tags:
        - Service Orders
      summary: Cria uma ordem de serviço identificando o cliente por CPF ou CNPJ
      operationId: createServiceOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceOrderRequest'
            examples:
              clienteVeiculoExistentes:
                summary: OS para cliente e veículo já cadastrados
                description: Use este exemplo depois de autenticar como admin. Ele já gera orçamento automaticamente.
                value:
                  customerDocument: '12345678909'
                  vehicleId: 20000000-0000-0000-0000-000000000001
                  diagnosticNotes: Cliente relata ruido ao frear e vibracao no pedal.
                  services:
                    - serviceId: 30000000-0000-0000-0000-000000000004
                      quantity: 1
                  parts:
                    - partId: 40000000-0000-0000-0000-000000000005
                      quantity: 1
                    - partId: 40000000-0000-0000-0000-000000000007
                      quantity: 1
                  generateBudget: true
              clienteVeiculoNovos:
                summary: OS cadastrando cliente e veículo no mesmo fluxo
                description: Demonstra identificação por CPF/CNPJ, cadastro de veículo e geração automática de orçamento.
                value:
                  customerDocument: '52998224725'
                  customer:
                    name: Yasmin Barcelos
                    phone: '11987654321'
                    email: yasmin.cliente@example.com
                    address:
                      street: Avenida Paulista
                      number: '1000'
                      complement: Sala 10
                      neighborhood: Bela Vista
                      city: Sao Paulo
                      state: SP
                      zipCode: 01310-100
                  vehicle:
                    plate: TST1A23
                    brand: Honda
                    model: Civic Touring
                    year: 2021
                    mileage: 42000
                  diagnosticNotes: Cliente solicita revisao preventiva antes de viagem.
                  services:
                    - serviceId: 30000000-0000-0000-0000-000000000003
                      quantity: 1
                  parts:
                    - partId: 40000000-0000-0000-0000-000000000001
                      quantity: 4
                    - partId: 40000000-0000-0000-0000-000000000002
                      quantity: 1
                  generateBudget: true
      responses:
        '201':
          description: Ordem de serviço criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags:
        - Service Orders
      summary: Lista ordens de serviço
      operationId: listServiceOrders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ServiceOrderStatus'
        - name: customerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: vehicleId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: createdFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Ordens de serviço encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/service-orders/tracking:
    get:
      tags:
        - Service Orders
      summary: Acompanha o progresso de ordens de serviço pelo cliente
      description: Consulte por id da OS ou por CPF/CNPJ do cliente junto com a placa do veículo.
      operationId: trackServiceOrders
      parameters:
        - name: serviceOrderId
          in: query
          required: false
          example: 50000000-0000-0000-0000-000000000002
          schema:
            type: string
            format: uuid
        - name: customerDocument
          in: query
          required: false
          example: '12345678909'
          schema:
            type: string
            minLength: 11
            maxLength: 18
            pattern: "^\\d{11}$|^\\d{14}$|^\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}$|^\\d{2}\\.\\d{3}\\.\\d{3}/\\d{4}-\\d{2}$"
        - name: plate
          in: query
          required: false
          example: MCA1D23
          schema:
            type: string
            minLength: 7
            maxLength: 8
            pattern: '^[A-Za-z]{3}-?[0-9]{4}$|^[A-Za-z]{3}-?[0-9][A-Za-z][0-9]{2}$'
      responses:
        '200':
          description: Progresso da ordem encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderTrackingListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}:
    get:
      tags:
        - Service Orders
      summary: Busca uma ordem de serviço por id
      operationId: getServiceOrderById
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      responses:
        '200':
          description: Ordem de serviço encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/services:
    post:
      tags:
        - Service Orders
      summary: Adiciona um serviço a uma ordem de serviço
      operationId: addServiceToServiceOrder
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddServiceOrderServiceRequest'
            examples:
              adicionarServico:
                summary: Inclui serviço solicitado na OS
                value:
                  serviceId: 30000000-0000-0000-0000-000000000005
                  quantity: 1
      responses:
        '200':
          description: Serviço adicionado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/parts:
    post:
      tags:
        - Service Orders
      summary: Adiciona uma peça a uma ordem de serviço
      operationId: addPartToServiceOrder
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddServiceOrderPartRequest'
            examples:
              adicionarPeca:
                summary: Inclui peça ou insumo na OS
                value:
                  partId: 40000000-0000-0000-0000-000000000005
                  quantity: 1
      responses:
        '200':
          description: Peça adicionada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/budget/generate:
    post:
      tags:
        - Service Orders
      summary: Gera o orçamento de uma ordem de serviço
      operationId: generateServiceOrderBudget
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      responses:
        '200':
          description: Orçamento gerado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/budget/approve:
    post:
      tags:
        - Service Orders
      summary: Aprova o orçamento de uma ordem de serviço
      operationId: approveServiceOrderBudget
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      responses:
        '200':
          description: Orçamento aprovado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/status:
    patch:
      tags:
        - Service Orders
      summary: Atualiza o status de uma ordem de serviço
      operationId: updateServiceOrderStatus
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceOrderStatusRequest'
            examples:
              iniciarDiagnostico:
                summary: Muda de Recebida para Em diagnóstico
                value:
                  status: IN_DIAGNOSIS
              iniciarExecucao:
                summary: Muda para Em execução após aprovação do orçamento
                value:
                  status: IN_PROGRESS
              finalizar:
                summary: Finaliza a execução da OS
                value:
                  status: FINISHED
              entregar:
                summary: Marca o veículo como entregue ao cliente
                value:
                  status: DELIVERED
      responses:
        '200':
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CustomerIdPathParam:
      name: customerId
      in: path
      required: true
      example: 10000000-0000-0000-0000-000000000001
      schema:
        type: string
        format: uuid
...
    ServiceOrderIdPathParam:
      name: serviceOrderId
      in: path
      required: true
      example: 50000000-0000-0000-0000-000000000002
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Requisição invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 400
            error: Bad Request
            message: Invalid request body
            path: /api/v1/service-orders
            details:
              - customerDocument: must not be blank
    Unauthorized:
      description: Credenciais ausentes ou invalidas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 401
            error: Unauthorized
            message: Missing or invalid credentials
            path: /api/v1/customers
            details: [ ]
...
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 404
            error: Not Found
            message: Resource not found
            path: /api/v1/customers/00000000-0000-0000-0000-000000000000
            details: [ ]
...
  schemas:
...
    Address:
      type: object
      required:
        - street
        - number
        - neighborhood
        - city
        - state
        - zipCode
      properties:
        street:
          type: string
          minLength: 2
          maxLength: 120
          example: Avenida Paulista
        number:
          type: string
          minLength: 1
          maxLength: 20
          example: '1000'
        complement:
          type: string
          maxLength: 80
          example: Sala 10
        neighborhood:
          type: string
          minLength: 2
          maxLength: 80
          example: Bela Vista
        city:
          type: string
          minLength: 2
          maxLength: 80
          example: Sao Paulo
        state:
          type: string
          minLength: 2
          maxLength: 2
          example: SP
        zipCode:
          type: string
          minLength: 8
          maxLength: 9
          pattern: "^\\d{5}-?\\d{3}$"
          example: 01310-100
...
    VehicleResponse:
      type: object
      required:
        - id
        - customerId
        - plate
        - brand
        - model
        - year
        - mileage
        - active
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        plate:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        active:
          type: boolean
...
    CreateServiceOrderRequest:
      type: object
      required:
        - customerDocument
        - diagnosticNotes
        - services
      properties:
        customerDocument:
          type: string
          minLength: 11
          maxLength: 18
          pattern: "^\\d{11}$|^\\d{14}$|^\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}$|^\\d{2}\\.\\d{3}\\.\\d{3}/\\d{4}-\\d{2}$"
          example: '12345678909'
          description: CPF ou CNPJ do cliente com ou sem formatação. Se o cliente nao existir, informe tambem customer.
        customer:
          $ref: '#/components/schemas/CreateServiceOrderCustomerRequest'
        vehicleId:
          type: string
          format: uuid
          description: >
            Identificador de veículo ja cadastrado.
            Se ausente, informe vehicle para cadastro ou vinculo por placa.
        vehicle:
          $ref: '#/components/schemas/CreateServiceOrderVehicleRequest'
        diagnosticNotes:
          type: string
          minLength: 5
          maxLength: 2000
        services:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateServiceOrderServiceRequest'
          description: Serviços solicitados para compor automáticamente o orçamento.
        parts:
          type: array
          items:
            $ref: '#/components/schemas/CreateServiceOrderPartRequest'
          description: Peças ou insumos necessários para compor automáticamente o orçamento.
        generateBudget:
          type: boolean
          default: true
          description: Quando true, gera o orçamento automáticamente e deixa a OS aguardando aprovação.
    CreateServiceOrderCustomerRequest:
      type: object
      required:
        - name
        - phone
        - email
        - address
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 10
          maxLength: 20
        email:
          type: string
          format: email
          maxLength: 120
        address:
          $ref: '#/components/schemas/Address'
    CreateServiceOrderVehicleRequest:
      type: object
      required:
        - plate
        - brand
        - model
        - year
      properties:
        plate:
          type: string
          minLength: 7
          maxLength: 8
          pattern: '^[A-Za-z]{3}-?[0-9]{4}$|^[A-Za-z]{3}-?[0-9][A-Za-z][0-9]{2}$'
          example: ABC1D23
        brand:
          type: string
          minLength: 2
          maxLength: 60
          example: Honda
        model:
          type: string
          minLength: 1
          maxLength: 80
          example: Civic
        year:
          type: integer
          minimum: 1900
          example: 2020
        mileage:
          type: integer
          minimum: 0
          default: 0
          example: 35000
    CreateServiceOrderServiceRequest:
      type: object
      required:
        - serviceId
        - quantity
      properties:
        serviceId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          default: 1
    CreateServiceOrderPartRequest:
      type: object
      required:
        - partId
        - quantity
      properties:
        partId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          default: 1
    AddServiceOrderServiceRequest:
      type: object
      required:
        - serviceId
        - quantity
      properties:
        serviceId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    AddServiceOrderPartRequest:
      type: object
      required:
        - partId
        - quantity
      properties:
        partId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    UpdateServiceOrderStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/ServiceOrderStatus'
    ServiceOrderServiceItem:
      type: object
      required:
        - serviceId
        - name
        - quantity
        - unitPrice
        - totalPrice
      properties:
        serviceId:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double
          minimum: 0
        totalPrice:
          type: number
          format: double
          minimum: 0
    ServiceOrderPartItem:
      type: object
      required:
        - partId
        - name
        - sku
        - quantity
        - unitPrice
        - totalPrice
      properties:
        partId:
          type: string
          format: uuid
        name:
          type: string
        sku:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double
          minimum: 0
        totalPrice:
          type: number
          format: double
          minimum: 0
    ServiceOrderResponse:
      type: object
      required:
        - id
        - customerId
        - vehicleId
        - status
        - diagnosticNotes
        - services
        - parts
        - servicesTotal
        - partsTotal
        - totalAmount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ServiceOrderStatus'
        diagnosticNotes:
          type: string
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderServiceItem'
        parts:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderPartItem'
        servicesTotal:
          type: number
          format: double
          minimum: 0
          description: Valor total dos serviços do orçamento.
        partsTotal:
          type: number
          format: double
          minimum: 0
          description: Valor total das peças e insumos do orçamento.
        totalAmount:
          type: number
          format: double
          minimum: 0
          description: Valor total geral do orçamento.
        createdAt:
          type: string
          format: date-time
        budgetGeneratedAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
    ServiceOrderListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderResponse'
    ServiceOrderTrackingListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderTrackingResponse'
    ServiceOrderTrackingResponse:
      type: object
      required:
        - id
        - customerId
        - vehicle
        - status
        - diagnosticNotes
        - services
        - parts
        - budget
        - statusHistory
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        vehicle:
          $ref: '#/components/schemas/VehicleResponse'
        status:
          $ref: '#/components/schemas/ServiceOrderTrackingStatus'
        diagnosticNotes:
          type: string
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderServiceItem'
        parts:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderPartItem'
        budget:
          $ref: '#/components/schemas/ServiceOrderBudgetTrackingResponse'
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderStatusHistoryItem'
        createdAt:
          type: string
          format: date-time
...
    ServiceOrderStatusHistoryItem:
      type: object
      required:
        - status
        - occurredAt
        - description
      properties:
        status:
          $ref: '#/components/schemas/ServiceOrderTrackingStatus'
        occurredAt:
          type: string
          format: date-time
        description:
          type: string
...
    ServiceOrderStatus:
      type: string
      enum:
        - RECEIVED
        - IN_DIAGNOSIS
        - WAITING_APPROVAL
        - IN_PROGRESS
        - FINISHED
        - DELIVERED
    ServiceOrderTrackingStatus:
      type: string
      enum:
        - RECEBIDA
        - EM_DIAGNOSTICO
        - AGUARDANDO_APROVACAO
        - EM_EXECUCAO
        - FINALIZADA
        - ENTREGUE
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          minimum: 100
          maximum: 599
        error:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        path:
          type: string
          minLength: 1
        details:
          type: array
          items:
            type: string